+++ src/aa_device_convolution.cu
@@ -18,7 +18,18 @@ namespace astroaccelerate {
 	}
 
 
-	void conv_OLS_customFFT(float2 *d_input_signal, float *d_output_plane, float2 *d_filters, int signal_length, int convolution_length, int useful_part_size, int offset, int nConvolutions, int nFilters, float scale){
+	void conv_OLS_customFFT(
+			float2 *d_input_signal, 
+			float *d_output_plane, 
+			float2 *d_filters, 
+			int64_t signal_length, 
+			int64_t convolution_length, 
+			int64_t useful_part_size, 
+			int64_t offset, 
+			int64_t nConvolutions, 
+			int64_t nFilters, 
+			float scale
+	){
 		dim3 gridSize(nConvolutions, 1, 1);
 		dim3 blockSize(convolution_length/4, 1, 1);
 		
@@ -26,9 +37,18 @@ namespace astroaccelerate {
 	}
 
 
-	void convolve_signal_C2C(float2 *d_input_signal, float *d_output_plane, float2 *d_filters, int signal_length, int nFilters, int filter_halfwidth, int convolution_length, float scale){
-		int useful_part_size = convolution_length - 2*filter_halfwidth + 1;
-		int nSegments        = (signal_length + useful_part_size - 1)/useful_part_size;
+	void convolve_signal_C2C(
+			float2 *d_input_signal, 
+			float *d_output_plane, 
+			float2 *d_filters, 
+			int64_t signal_length, 
+			int64_t nFilters, 
+			int64_t filter_halfwidth, 
+			int64_t convolution_length, 
+			float scale
+	){
+		int64_t useful_part_size = convolution_length - 2*filter_halfwidth + 1;
+		int64_t nSegments        = (signal_length + useful_part_size - 1)/useful_part_size;
 		conv_OLS_customFFT(d_input_signal, d_output_plane, d_filters, signal_length, convolution_length, useful_part_size, filter_halfwidth, nSegments, nFilters, scale);
 	}
